Js Lint error fixed :wrench:, Fixes #110
