+++ app/main/index.js
@@ -9,11 +9,11 @@ const {dialog} = require('electron');
 const https = require('https');
 const http = require('http');
 const electronLocalshortcut = require('electron-localshortcut');
-// const Configstore = require('configstore');
+// Not using // const Configstore = require('configstore');
 const Configstore = require('electron-config');
 const JsonDB = require('node-json-db');
 const isDev = require('electron-is-dev');
-const tray = require('./tray');
+// Not using now //const tray = require('./tray');
 const appMenu = require('./menu');
 const {linkIsInternal, skipImages} = require('./link-helper');
 const {appUpdater} = require('./autoupdater');
@@ -21,12 +21,11 @@ const {appUpdater} = require('./autoupdater');
 const db = new JsonDB(app.getPath('userData') + '/domain.json', true, true);
 const data = db.getData('/');
 
-// adds debug features like hotkeys for triggering dev tools and reload
+// Adds debug features like hotkeys for triggering dev tools and reload
 require('electron-debug')();
 
-// const conf = new Configstore('Zulip-Desktop');
+// Const conf = new Configstore('Zulip-Desktop');
 const conf = new Configstore();
-
 
 function userOS() {
 	if (os.platform() === 'darwin') {
@@ -44,10 +43,10 @@ function userOS() {
 	}
 }
 
-// setting userAgent so that server-side code can identify the desktop app
+// Setting userAgent so that server-side code can identify the desktop app
 const isUserAgent = 'ZulipElectron/' + app.getVersion() + ' ' + userOS();
 
-// prevent window being garbage collected
+// Prevent window being garbage collected
 let mainWindow;
 let targetLink;
 
@@ -162,8 +161,7 @@ function updateDockBadge(title) {
 	if (process.platform === 'darwin') {
 		app.setBadgeCount(messageCount);
 	}
-         mainWindow.webContents.send('tray', messageCount);
-
+	mainWindow.webContents.send('tray', messageCount);
 }
 
 function createMainWindow() {
@@ -223,7 +221,7 @@ function createMainWindow() {
 		});
 	});
 
-	// on osx it's 'moved'
+	// On osx it's 'moved'
 	win.on('move', function () {
 		const pos = this.getPosition();
 		conf.set({
@@ -232,17 +230,18 @@ function createMainWindow() {
 		});
 	});
 
-	// stop page to update it's title
+	// Stop page to update it's title
 	win.on('page-title-updated', (e, title) => {
 		e.preventDefault();
 		updateDockBadge(title);
 	});
 
-         //To destroy tray icon when navigate to a new URL
-	win.webContents.on('will-navigate', (e, url) => {
-		// console.log(url);
-		win.webContents.send('destroytray');
-   	});
+         //  To destroy tray icon when navigate to a new URL
+	win.webContents.on('will-navigate', e => {
+		if (e) {
+			win.webContents.send('destroytray');
+		}
+	});
 
 	return win;
 }
@@ -251,7 +250,7 @@ function createMainWindow() {
 app.commandLine.appendSwitch('ignore-certificate-errors', 'true');
 
 app.on('window-all-closed', () => {
-	// unregister all the shortcuts so that they don't interfare with other apps
+	// Unregister all the shortcuts so that they don't interfare with other apps
 	electronLocalshortcut.unregisterAll(mainWindow);
 	if (process.platform !== 'darwin') {
 		app.quit();
@@ -267,7 +266,7 @@ app.on('activate', () => {
 app.on('ready', () => {
 	electron.Menu.setApplicationMenu(appMenu);
 	mainWindow = createMainWindow();
-	//tray.create();
+	// Not using for now // tray.create();
 
 	const page = mainWindow.webContents;
 
@@ -314,12 +313,12 @@ app.on('ready', () => {
 });
 
 app.on('will-quit', () => {
-	// unregister all the shortcuts so that they don't interfare with other apps
+	// Unregister all the shortcuts so that they don't interfare with other apps
 	electronLocalshortcut.unregisterAll(mainWindow);
 });
 
 ipc.on('new-domain', (e, domain) => {
-	// mainWindow.loadURL(domain);
+	// MainWindow.loadURL(domain);
 	if (!mainWindow) {
 		mainWindow = createMainWindow();
 		mainWindow.loadURL(domain);
